20.10.2 release

+++ htdocs/webservice/admin/oauthv1sregister.php
@@ -137,10 +137,12 @@ function webservices_server_submit(Pieform $form, $values) {
                                 DELETE FROM {oauth_server_token}
                                 WHERE osr_id_ref = ?
                                 ', array($dbserver->id));
-            delete_records_sql('
-                                DELETE FROM {lti_assessment}
-                                WHERE oauthserver = ?
-                                ', array($dbserver->id));
+            if (db_table_exists('lti_assessment')) {
+                delete_records_sql('
+                                    DELETE FROM {lti_assessment}
+                                    WHERE oauthserver = ?
+                                    ', array($dbserver->id));
+            }
             $store->deleteServer($dbserver->consumer_key, $dbserver->userid, $is_admin);
             $SESSION->add_ok_msg(get_string('oauthserverdeleted', 'auth.webservice'));
         }
